requirements.txt
networkx
matplotlib
pyvis
playwright
readability-lxml
bs4
asyncio
